address git comments and port over ethereal/kinetic elite equipments
