--- SS2-Project/Assets/Starstorm2/Modules/Pickups/Equipments/Unfinished/EliteKineticEquipment.cs
@@ -1,23 +1,0 @@
-﻿using RoR2;
-using System.Collections.Generic;
-namespace SS2.Equipments
-{
-#if DEBUG
-    /*
-    public sealed class EliteKineticEquipment : SS2EliteEquipment
-    {
-        public override EquipmentDef EquipmentDef { get; } = SS2Assets.LoadAsset<EquipmentDef>("EliteKineticEquipment", SS2Bundle.Indev);
-
-        public override List<MSEliteDef> EliteDefs => new List<MSEliteDef>
-        {
-            SS2Assets.LoadAsset<MSEliteDef>("edKinetic", SS2Bundle.Indev),
-            SS2Assets.LoadAsset<MSEliteDef>("edKineticHonor", SS2Bundle.Indev)
-        };
-
-        public override bool FireAction(EquipmentSlot slot)
-        {
-            return false;
-        }
-    }*/
-#endif
-}
